% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusignrank.R
\name{clusignrank}
\alias{clusignrank}
\title{The Wilcoxon Signed Rank Test for Paired Comparisons of Clustered Data}
\usage{
clusignrank(z, id = NULL, data = parent.frame())
}
\arguments{
\item{z}{The name of score for each observation
in the data}

\item{id}{The name of id of cluster for each
observation in the data. If not specified, each score has
its own id, which reduces to non-clustereddata.}

\item{data}{An optional data frame, list or environment (or object
coercible by \code{as.data.frame} to a data frame) containing the
variables in the model.  If not found in \code{data}, the
variables are taken from \code{environment(formula)}, typically
the environment from which this function is called.}
}
\value{
a list with the following components
\item{T_c}{Clustered Wilcoxon signed ranksum statistic}
\item{Var_t}{Variance of clustered Wilcoxon signed ranksum statistic}
\item{W_c}{Standardized clustered Wilcoxon signed ranksum statistic}
\item{p.value}{P-value for clustered Wilcoxon ranksum test statistic W_c}
\item{n}{Total number of observations}
\item{m}{Total number of clusters}
}
\description{
Signed rank test for clustered data where there are for example
change score for each of 1 or more subjects per cluster and
one wants to test the hypothesis that the change score
distribution is symmetric about zero using the Wilcoxon signed
rank test using the subunit as the unit of change.
}
\note{
The function will determine if the dataset is balanced or
unbalanced, and treat them accordingly
Zero values can be included in a dataset but are not used in
the analysis.
}
\examples{
data(crsd)
clusignrank(z, id, data = crsd)
data(crsd.unb)
clusignrank(z, id, data = crsd.unb)
}
\references{
Bernard Rosner, Robert J. Glynn, Mei-Ling Ting Lee(2006)
\emph{The Wilcoxon Signed Rank Test for Paired Comparisons of
Clustered Data.} Biometrics, \bold{62}, 185-192.
}

