\name{crsd.unb}
\alias{crsd.unb}
\docType{data}
\title{
Unbalanced clustered data for test of function clusignrank
}
\description{
Test unbalanced clustered data include 2 variables: id, z. 
}
\usage{data("crsd.unb")}
\format{
  A data frame with 748 observations on the following 2 variables.
  \describe{
    \item{\code{id}}{id of cluster}
    \item{\code{z}}{score of observation}
  }
}

\source{
URL: https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/cluster-signed-rank-test-for-balancedunbalanced-design/sample-data-for-unbalanced-data-1
}

\examples{
data(crsd.unb)
}
\keyword{datasets}
